modified query

+++ migrations/20240506070933_create_products_table.up.sql
@@ -5,7 +5,7 @@ CREATE TABLE products (
   category varchar NOT NULL,
   image_urls text[] NOT NULL,
   notes varchar NOT NULL,
-  price int NOT NULL,
+  price numeric NOT NULL,
   stock int NOT NULL,
   location varchar NOT NULL,
   is_available bool NOT NULL,
